requirements.txt
python-telegram-bot==13.3
PyMySQL==1.0.2
